+++ test/carbonite/transaction_test.exs
@@ -1,9 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 defmodule Carbonite.TransactionTest do
-  use ExUnit.Case, async: true
+  use Carbonite.APICase, async: true
   import Carbonite.Transaction
-  import Ecto.Changeset
 
   describe "changeset/1" do
     test "transaction_changesets an Ecto.Changeset for a transaction" do
@@ -19,8 +18,8 @@ defmodule Carbonite.TransactionTest do
     end
 
     test "merges metadata from process dictionary" do
-      put_meta(:foo, 1)
-      put_meta(:bar, 1)
+      Carbonite.Transaction.put_meta(:foo, 1)
+      Carbonite.Transaction.put_meta(:bar, 1)
       %Ecto.Changeset{} = changeset = changeset(%{meta: %{foo: 2}})
 
       assert get_field(changeset, :meta) == %{foo: 2, bar: 1}